requirements.txt
httpx[http2]==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
python-dotenv==1.0.1
tenacity==8.3.0
tqdm==4.66.4
charset-normalizer==3.3.2
google-genai
